ajustes em aluno

+++ GeradorCertificados2/Models/AlunoViewModel.cs
@@ -4,5 +4,5 @@ public class Aluno
 {
     public int Id { get; set; }
     public string Nome { get; set; }
-    public long Matricula { get; set; }
+    public int Matricula { get; set; }
 }
